+++ bootique-jersey-client-instrumented/src/main/java/io/bootique/jersey/client/instrumented/threshold/JerseyClientHealthChecks.java
@@ -5,10 +5,10 @@ import io.bootique.metrics.health.HealthCheckGroup;
 
 import java.util.Map;
 
-public class JerseyHealthChecks implements HealthCheckGroup {
+public class JerseyClientHealthChecks implements HealthCheckGroup {
     private final Map<String, HealthCheck> healthChecks;
 
-    public JerseyHealthChecks(Map<String, HealthCheck> healthChecks) {
+    public JerseyClientHealthChecks(Map<String, HealthCheck> healthChecks) {
         this.healthChecks = healthChecks;
     }