+++ bootique-jersey-client-instrumented/src/main/java/io/bootique/jersey/client/instrumented/threshold/JerseyClientHealthChecksFactory.java
@@ -15,46 +15,46 @@ import java.util.Map;
 import java.util.function.Supplier;
 
 @BQConfig
-public class ThresholdHealthCheckFactory {
+public class JerseyClientHealthChecksFactory {
 
-    public static final String THRESHOLD_REQUESTS_CHECK =  JerseyClientInstrumentedModule
+    public static final String REQUESTS_RATE_CHECK =  JerseyClientInstrumentedModule
             .METRIC_NAMING
-            .name("Threshold", "Requests");
+            .name("Requests", "Rate");
 
-    private DoubleRangeFactory timeRequestsThresholds;
+    private DoubleRangeFactory requestsPerMin;
 
     @BQConfigProperty
-    public void setTimeRequestsThresholds(DoubleRangeFactory timeRequestsThresholds) {
-        this.timeRequestsThresholds = timeRequestsThresholds;
+    public void setRequestsPerMin(DoubleRangeFactory requestsPerMin) {
+        this.requestsPerMin = requestsPerMin;
     }
 
-    public JerseyHealthChecks createThresholdHealthCheck(MetricRegistry metricRegistry) {
-        return new JerseyHealthChecks(createHealthChecksMap(metricRegistry));
+    public JerseyClientHealthChecks createHealthChecks(MetricRegistry metricRegistry) {
+        return new JerseyClientHealthChecks(createHealthChecksMap(metricRegistry));
     }
 
     protected Map<String, HealthCheck> createHealthChecksMap(MetricRegistry registry) {
         Map<String, HealthCheck> checks = new HashMap<>(3);
-        checks.put(THRESHOLD_REQUESTS_CHECK, createTimeRequestsCheck(registry));
+        checks.put(REQUESTS_RATE_CHECK, createTimeRequestsCheck(registry));
         return checks;
     }
 
     private HealthCheck createTimeRequestsCheck(MetricRegistry registry) {
-        ValueRange<Double> range = getTimeRequestsThresholds();
+        ValueRange<Double> range = getRequestsPerMin();
         Supplier<Double> deferredGauge = ()
                 -> registry.timer(RequestTimer.TIMER_NAME).getOneMinuteRate();
 
         return new ValueRangeCheck<>(range, deferredGauge);
     }
 
-    protected ValueRange<Double> getTimeRequestsThresholds() {
+    protected ValueRange<Double> getRequestsPerMin() {
 
         // init min if it wasn't set...
-        if (timeRequestsThresholds != null) {
-            if (timeRequestsThresholds.getMin() == null) {
-                timeRequestsThresholds.setMin(0);
+        if (requestsPerMin != null) {
+            if (requestsPerMin.getMin() == null) {
+                requestsPerMin.setMin(0);
             }
 
-            return timeRequestsThresholds.createRange();
+            return requestsPerMin.createRange();
         }
 
         return ValueRange.builder(Double.class).min(0.0).warning(3.0).critical(15.0).build();